Refactor user deletion methods to exclude admin users and add endpoint to delete all non-admin users

+++ src/main/java/com/rzodeczko/infrastructure/persistence/repository/MongoUserRepository.java
@@ -1,5 +1,6 @@
 package com.rzodeczko.infrastructure.persistence.repository;
 
+import com.rzodeczko.application.security.enums.Role;
 import com.rzodeczko.infrastructure.persistence.document.UserDocument;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import reactor.core.publisher.Mono;
@@ -11,6 +12,7 @@ public interface MongoUserRepository extends ReactiveMongoRepository<UserDocumen
 
     default Mono<UserDocument> deleteByUsername(String username) {
         return findByUsername(username)
+                .filter(user -> user.getRole() != Role.ROLE_ADMIN)
                 .flatMap(userDocument -> deleteById(userDocument.getId()).thenReturn(userDocument));
     }
 }
